campusconnect_backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
Pillow>=10.0.0

